+++ androidutilities-room/src/test/kotlin/com/heyzeusv/androidutitlities/room/RoomBackupRestoreCreatorTest.kt
@@ -30,7 +30,7 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
             )
         )
         assertEquals(KotlinCompilation.ExitCode.OK, kspCompileResult.result.exitCode)
-        assertEquals(6, kspCompileResult.generatedFiles.size)
+        assertEquals(7, kspCompileResult.generatedFiles.size)
         kspCompileResult.assertFileEquals(expectedRoomBackupRestore, "RoomBackupRestore.kt")
     }
 
@@ -51,7 +51,7 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
             kspArguments = mutableMapOf("roomUtilHilt" to "TRUE"),
         )
         assertEquals(KotlinCompilation.ExitCode.OK, kspCompileResult.result.exitCode)
-        assertEquals(6, kspCompileResult.generatedFiles.size)
+        assertEquals(7, kspCompileResult.generatedFiles.size)
         kspCompileResult.assertFileEquals(
             expectedRoomBackupRestoreWithHiltOptionValue,
             "RoomBackupRestore.kt",
@@ -76,7 +76,7 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
             kspArguments = mutableMapOf("roomUtilDb" to "FALSE"),
         )
         assertEquals(KotlinCompilation.ExitCode.OK, kspCompileResult.result.exitCode)
-        assertEquals(5, kspCompileResult.generatedFiles.size)
+        assertEquals(6, kspCompileResult.generatedFiles.size)
         assertFalse(kspCompileResult.generatedFiles.any { it.name == "RoomBackupRestore.kt" })
     }
 
@@ -96,6 +96,9 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
             import kotlin.Boolean
             import kotlin.String
             import kotlin.Unit
+            import test.RoomUtilStatus.Error
+            import test.RoomUtilStatus.Progress
+            import test.RoomUtilStatus.Success
 
             public class RoomBackupRestore(
               private val context: Context,
@@ -106,61 +109,93 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
                 // include file extension to dbFileName!!
               }
               public fun backup(appBackupDirectoryUri: Uri) {
+                _status.value = Progress(R.string.backup_progress_started)
                 val appBackupDirectory = DocumentFile.fromTreeUri(context, appBackupDirectoryUri)!!
                 if (!appBackupDirectory.exists()) {
-                  return // directory not found
+                  _status.value = Error(R.string.backup_error_missing_directory)
+                  return
                 }
 
-                val newBackupDirectory = createNewDirectory(appBackupDirectory) ?:
-                  return // new directory could not be created
+                val newBackupDirectory = createNewDirectory(appBackupDirectory)
+                if (newBackupDirectory == null) {
+                  _status.value = Error(R.string.backup_error_create_directory_failed)
+                  return
+                }
                 val dbPath = context.getDatabasePath(dbFileName).path
                 val dbFile = DocumentFile.fromFile(File(dbPath))
                 val dbWalFile = DocumentFile.fromFile(File(""${'"'}${'$'}dbPath-wal""${'"'}))
                 val dbShmFile = DocumentFile.fromFile(File(""${'"'}${'$'}dbPath-shm""${'"'}))
-                if (!dbFile.exists()) return // main db file not found
+                if (!dbFile.exists()) {
+                  _status.value = Error(R.string.backup_error_missing_file, dbPath)
+                  return
+                }
 
                 val newFiles = mutableListOf(newBackupDirectory)
-                val bkpDbFile = newBackupDirectory.createFile("text/*", dbFileName) ?:
-                  return // error creating backup file
+                val bkpDbFile = newBackupDirectory.createFile("text/*", dbFileName)
+                if (bkpDbFile == null) {
+                  _status.value = Error(R.string.backup_error_create_file_failed, dbFileName)
+                  return
+                }
                 var dbFileCopyStatus = dbFile.copyTo(bkpDbFile)
                 if (!dbFileCopyStatus) {
                   bkpDbFile.delete()
                   newBackupDirectory.delete()
-                  return // failed to copy main db file
+                  _status.value = Error(R.string.backup_error_failed, dbFileName)
+                  return
                 }
                 newFiles.add(bkpDbFile)
+                _status.value = Progress(R.string.backup_progress_file_success, dbFileName)
 
                 if (dbWalFile.exists()) {
-                  val bkpDbWalFile = newBackupDirectory.createFile("text/*", ""${'"'}${'$'}dbFileName-wal""${'"'}) ?:
-                    return // error creating backup file
+                  val bkpDbWalFile = newBackupDirectory.createFile("text/*", ""${'"'}${'$'}dbFileName-wal""${'"'})
+                  if (bkpDbWalFile == null) {
+                    _status.value = Error(R.string.backup_error_create_file_failed, ""${'"'}${'$'}dbFileName-wal""${'"'})
+                    return
+                  }  
                   dbFileCopyStatus = dbWalFile.copyTo(bkpDbWalFile)
                   if (!dbFileCopyStatus) {
                     newFiles.forEach { it.delete() }
                     bkpDbWalFile.delete()
-                    return // failed to copy wal file
+                    newBackupDirectory.delete()
+                    _status.value = Error(R.string.backup_error_failed, ""${'"'}${'$'}dbFileName-wal""${'"'})
+                    return
                   }
                   newFiles.add(bkpDbWalFile)
+                  _status.value = Progress(R.string.backup_progress_file_success, ""${'"'}${'$'}dbFileName-wal""${'"'})
                 }
                     
                 if (dbShmFile.exists()) {
-                  val bkpDbShmFile = newBackupDirectory.createFile("text/*", ""${'"'}${'$'}dbFileName-shm""${'"'}) ?:
-                    return // error creating backup file
+                  val bkpDbShmFile = newBackupDirectory.createFile("text/*", ""${'"'}${'$'}dbFileName-shm""${'"'})
+                  if (bkpDbShmFile == null) {
+                    _status.value = Error(R.string.backup_error_create_file_failed, ""${'"'}${'$'}dbFileName-shm""${'"'})
+                    return
+                  } 
                   dbFileCopyStatus = dbShmFile.copyTo(bkpDbShmFile)
                   if (!dbFileCopyStatus) {
                     newFiles.forEach { it.delete() }
                     bkpDbShmFile.delete()
-                    return // failed to copy shm file
+                    newBackupDirectory.delete()
+                    _status.value = Error(R.string.backup_error_failed, ""${'"'}${'$'}dbFileName-shm""${'"'})
+                    return
                   }
+                  _status.value = Progress(R.string.backup_progress_file_success, ""${'"'}${'$'}dbFileName-shm""${'"'})
                 }
+                _status.value = Success(R.string.backup_success)
               }
 
               public fun restore(selectedDirectoryUri: Uri, restartApp: () -> Unit = { restartAppStandard() }) {
+                _status.value = Progress(R.string.restore_progress_started)
                 val selectedDirectory = DocumentFile.fromTreeUri(context, selectedDirectoryUri)!!
                 if (!selectedDirectory.exists()) {
-                  return // directory doesn't exist
+                  _status.value = Error(R.string.restore_error_missing_directory)
+                  return
                 }
 
-                val bkpDbFile = selectedDirectory.findFile(dbFileName) ?: return // main db file not found
+                val bkpDbFile = selectedDirectory.findFile(dbFileName)
+                if (bkpDbFile == null) {
+                  _status.value = Error(R.string.restore_error_missing_db_file)
+                  return
+                }
                 val bkpDbWalFile = selectedDirectory.findFile(""${'"'}${'$'}dbFileName-wal""${'"'})
                 val bkpDbShmFile = selectedDirectory.findFile(""${'"'}${'$'}dbFileName-shm""${'"'})
 
@@ -168,11 +203,16 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
                 val dbFile = DocumentFile.fromFile(File(dbPath))
                 val dbWalFile = DocumentFile.fromFile(File(""${'"'}${'$'}dbPath-wal""${'"'}))
                 val dbShmFile = DocumentFile.fromFile(File(""${'"'}${'$'}dbPath-shm""${'"'}))
-                if (!dbFile.exists()) return // main db file not found
+                if (!dbFile.exists()) {
+                  _status.value = Error(R.string.restore_error_missing_db_file)
+                  return
+                }
 
                 // delete any existing content before restoring
                 PrintWriter(File(dbPath)).close()
                 bkpDbFile.copyTo(dbFile)
+                _status.value = Progress(R.string.restore_progress_file_success, dbFile.name!!)
+
                 // file doesn't exist in backup so delete current
                 if (bkpDbWalFile == null) {
                   dbWalFile.delete()
@@ -180,6 +220,7 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
                   // delete any existing content before restoring
                   PrintWriter(File(""${'"'}${'$'}dbPath-wal""${'"'})).close()
                   bkpDbWalFile.copyTo(dbWalFile)
+                  _status.value = Progress(R.string.restore_progress_file_success, dbWalFile.name!!)
                 }
                 // file doesn't exist in backup so delete current
                 if (bkpDbShmFile == null) {
@@ -188,8 +229,10 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
                   // delete any existing content before restoring
                   PrintWriter(File(""${'"'}${'$'}dbPath-shm""${'"'})).close()
                   bkpDbShmFile.copyTo(dbShmFile)
+                  _status.value = Progress(R.string.restore_progress_file_success, dbShmFile.name!!)
                 }
 
+                _status.value = Success(R.string.restore_success)
                 restartApp()
               }
 
@@ -250,6 +293,9 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
             import kotlin.Boolean
             import kotlin.String
             import kotlin.Unit
+            import test.RoomUtilStatus.Error
+            import test.RoomUtilStatus.Progress
+            import test.RoomUtilStatus.Success
 
             public class RoomBackupRestore @Inject constructor(
               private val context: Context,
@@ -260,61 +306,93 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
                 // include file extension to dbFileName!!
               }
               public fun backup(appBackupDirectoryUri: Uri) {
+                _status.value = Progress(R.string.backup_progress_started)
                 val appBackupDirectory = DocumentFile.fromTreeUri(context, appBackupDirectoryUri)!!
                 if (!appBackupDirectory.exists()) {
-                  return // directory not found
+                  _status.value = Error(R.string.backup_error_missing_directory)
+                  return
                 }
 
-                val newBackupDirectory = createNewDirectory(appBackupDirectory) ?:
-                  return // new directory could not be created
+                val newBackupDirectory = createNewDirectory(appBackupDirectory)
+                if (newBackupDirectory == null) {
+                  _status.value = Error(R.string.backup_error_create_directory_failed)
+                  return
+                }
                 val dbPath = context.getDatabasePath(dbFileName).path
                 val dbFile = DocumentFile.fromFile(File(dbPath))
                 val dbWalFile = DocumentFile.fromFile(File(""${'"'}${'$'}dbPath-wal""${'"'}))
                 val dbShmFile = DocumentFile.fromFile(File(""${'"'}${'$'}dbPath-shm""${'"'}))
-                if (!dbFile.exists()) return // main db file not found
+                if (!dbFile.exists()) {
+                  _status.value = Error(R.string.backup_error_missing_file, dbPath)
+                  return
+                }
 
                 val newFiles = mutableListOf(newBackupDirectory)
-                val bkpDbFile = newBackupDirectory.createFile("text/*", dbFileName) ?:
-                  return // error creating backup file
+                val bkpDbFile = newBackupDirectory.createFile("text/*", dbFileName)
+                if (bkpDbFile == null) {
+                  _status.value = Error(R.string.backup_error_create_file_failed, dbFileName)
+                  return
+                }
                 var dbFileCopyStatus = dbFile.copyTo(bkpDbFile)
                 if (!dbFileCopyStatus) {
                   bkpDbFile.delete()
                   newBackupDirectory.delete()
-                  return // failed to copy main db file
+                  _status.value = Error(R.string.backup_error_failed, dbFileName)
+                  return
                 }
                 newFiles.add(bkpDbFile)
+                _status.value = Progress(R.string.backup_progress_file_success, dbFileName)
 
                 if (dbWalFile.exists()) {
-                  val bkpDbWalFile = newBackupDirectory.createFile("text/*", ""${'"'}${'$'}dbFileName-wal""${'"'}) ?:
-                    return // error creating backup file
+                  val bkpDbWalFile = newBackupDirectory.createFile("text/*", ""${'"'}${'$'}dbFileName-wal""${'"'})
+                  if (bkpDbWalFile == null) {
+                    _status.value = Error(R.string.backup_error_create_file_failed, ""${'"'}${'$'}dbFileName-wal""${'"'})
+                    return
+                  }  
                   dbFileCopyStatus = dbWalFile.copyTo(bkpDbWalFile)
                   if (!dbFileCopyStatus) {
                     newFiles.forEach { it.delete() }
                     bkpDbWalFile.delete()
-                    return // failed to copy wal file
+                    newBackupDirectory.delete()
+                    _status.value = Error(R.string.backup_error_failed, ""${'"'}${'$'}dbFileName-wal""${'"'})
+                    return
                   }
                   newFiles.add(bkpDbWalFile)
+                  _status.value = Progress(R.string.backup_progress_file_success, ""${'"'}${'$'}dbFileName-wal""${'"'})
                 }
                     
                 if (dbShmFile.exists()) {
-                  val bkpDbShmFile = newBackupDirectory.createFile("text/*", ""${'"'}${'$'}dbFileName-shm""${'"'}) ?:
-                    return // error creating backup file
+                  val bkpDbShmFile = newBackupDirectory.createFile("text/*", ""${'"'}${'$'}dbFileName-shm""${'"'})
+                  if (bkpDbShmFile == null) {
+                    _status.value = Error(R.string.backup_error_create_file_failed, ""${'"'}${'$'}dbFileName-shm""${'"'})
+                    return
+                  } 
                   dbFileCopyStatus = dbShmFile.copyTo(bkpDbShmFile)
                   if (!dbFileCopyStatus) {
                     newFiles.forEach { it.delete() }
                     bkpDbShmFile.delete()
-                    return // failed to copy shm file
+                    newBackupDirectory.delete()
+                    _status.value = Error(R.string.backup_error_failed, ""${'"'}${'$'}dbFileName-shm""${'"'})
+                    return
                   }
+                  _status.value = Progress(R.string.backup_progress_file_success, ""${'"'}${'$'}dbFileName-shm""${'"'})
                 }
+                _status.value = Success(R.string.backup_success)
               }
 
               public fun restore(selectedDirectoryUri: Uri, restartApp: () -> Unit = { restartAppStandard() }) {
+                _status.value = Progress(R.string.restore_progress_started)
                 val selectedDirectory = DocumentFile.fromTreeUri(context, selectedDirectoryUri)!!
                 if (!selectedDirectory.exists()) {
-                  return // directory doesn't exist
+                  _status.value = Error(R.string.restore_error_missing_directory)
+                  return
                 }
 
-                val bkpDbFile = selectedDirectory.findFile(dbFileName) ?: return // main db file not found
+                val bkpDbFile = selectedDirectory.findFile(dbFileName)
+                if (bkpDbFile == null) {
+                  _status.value = Error(R.string.restore_error_missing_db_file)
+                  return
+                }
                 val bkpDbWalFile = selectedDirectory.findFile(""${'"'}${'$'}dbFileName-wal""${'"'})
                 val bkpDbShmFile = selectedDirectory.findFile(""${'"'}${'$'}dbFileName-shm""${'"'})
 
@@ -322,11 +400,16 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
                 val dbFile = DocumentFile.fromFile(File(dbPath))
                 val dbWalFile = DocumentFile.fromFile(File(""${'"'}${'$'}dbPath-wal""${'"'}))
                 val dbShmFile = DocumentFile.fromFile(File(""${'"'}${'$'}dbPath-shm""${'"'}))
-                if (!dbFile.exists()) return // main db file not found
+                if (!dbFile.exists()) {
+                  _status.value = Error(R.string.restore_error_missing_db_file)
+                  return
+                }
 
                 // delete any existing content before restoring
                 PrintWriter(File(dbPath)).close()
                 bkpDbFile.copyTo(dbFile)
+                _status.value = Progress(R.string.restore_progress_file_success, dbFile.name!!)
+
                 // file doesn't exist in backup so delete current
                 if (bkpDbWalFile == null) {
                   dbWalFile.delete()
@@ -334,6 +417,7 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
                   // delete any existing content before restoring
                   PrintWriter(File(""${'"'}${'$'}dbPath-wal""${'"'})).close()
                   bkpDbWalFile.copyTo(dbWalFile)
+                  _status.value = Progress(R.string.restore_progress_file_success, dbWalFile.name!!)
                 }
                 // file doesn't exist in backup so delete current
                 if (bkpDbShmFile == null) {
@@ -342,8 +426,10 @@ class RoomBackupRestoreCreatorTest : CreatorTestBase() {
                   // delete any existing content before restoring
                   PrintWriter(File(""${'"'}${'$'}dbPath-shm""${'"'})).close()
                   bkpDbShmFile.copyTo(dbShmFile)
+                  _status.value = Progress(R.string.restore_progress_file_success, dbShmFile.name!!)
                 }
 
+                _status.value = Success(R.string.restore_success)
                 restartApp()
               }
